requirements.txt
mcp>=1.0.0
dremio-client>=0.1.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0
pyarrow[flight]>=12.0.0
sqlalchemy-dremio>=2.1.0
openai>=1.0.0
flask>=2.3.0
flask-cors>=4.0.0
